+++ src/Braintree/RangeNode.cs
@@ -15,18 +15,26 @@ namespace Braintree
         }
 
         public T GreaterThanOrEqualTo(object min) {
-            Parent.AddRangeCriteria(Name, new SearchCriteria("min", min.ToString()));
+            Parent.AddRangeCriteria(Name, new SearchCriteria("min", NormalizedStringValueOfObject(min)));
             return Parent;
         }
 
         public T Is(object value) {
-            Parent.AddCriteria(Name, new SearchCriteria("is", value.ToString()));
+            Parent.AddCriteria(Name, new SearchCriteria("is", NormalizedStringValueOfObject(value)));
             return Parent;
         }
 
         public T LessThanOrEqualTo(object max) {
-            Parent.AddRangeCriteria(Name, new SearchCriteria("max", max.ToString()));
+            Parent.AddRangeCriteria(Name, new SearchCriteria("max", NormalizedStringValueOfObject(max)));
             return Parent;
+        }
+
+        internal static string NormalizedStringValueOfObject(object value) {
+            if (value is decimal) {
+                return ((decimal) value).ToString(System.Globalization.CultureInfo.InvariantCulture);
+            } else {
+                return value.ToString();
+            }
         }
     }
 }